+++ packages/insomnia/src/models/websocket-request.ts
@@ -52,8 +52,6 @@ export const init = (): BaseWebSocketRequest => ({
   description: '',
 });
 
-export const migrate = (doc: WebSocketRequest) => doc;
-
 export const create = (patch: Partial<WebSocketRequest> = {}) => {
   if (!patch.parentId) {
     throw new Error(`New WebSocketRequest missing \`parentId\`: ${JSON.stringify(patch)}`);
